Lots of work on the map. Almost happy with it.

+++ src/common/micdoodle8/mods/galacticraft/API/IPlanetSlotRenderer.java
@@ -8,5 +8,5 @@ public interface IPlanetSlotRenderer
 	
 	public String getPlanetName();
 	
-	public void renderSlot(int index, int x, int y, int slotHeight, Tessellator tessellator);
+	public void renderSlot(int index, int x, int y, float slotHeight, Tessellator tessellator);
 }
