+++ src/minecraft/micdoodle8/mods/galacticraft/titan/client/GCTitanSlotRenderer.java
@@ -20,7 +20,7 @@ public class GCTitanSlotRenderer implements IPlanetSlotRenderer
 	}
 
 	@Override
-	public void renderSlot(int index, int x, int y, int slotHeight, Tessellator tessellator) 
+	public void renderSlot(int index, int x, int y, float slotHeight, Tessellator tessellator) 
 	{
 		tessellator.startDrawingQuads();
 	    tessellator.addVertexWithUV(x - 5 - slotHeight * 1.3, 	y - 2 + slotHeight * 1.3, 	-90.0D, 0.35D, 0.65D);
